feat: add `CreateNewListModalForManageArtwork`
